+++ test/cases/exception-evaluate-template/webpack.config.js
@@ -19,7 +19,7 @@ module.exports = {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          method: 'compile',
+          method: 'render',
         },
       },