+++ test/cases/html-webpack-plugin-options/webpack.config.js
@@ -45,7 +45,7 @@ module.exports = {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          method: 'render',
+          //method: 'render',
         },
       },