+++ test/cases/include-alias-resolve.alias/webpack.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
-const PugPlugin = require('pug-plugin');
+const PugPlugin = require('../../pug-plugin');
+
 const basePath = path.resolve(__dirname);
 const webRootPath = path.join(__dirname, 'public/');