+++ test/cases/javascript-assets-compile/src/index.js
@@ -1,2 +1,10 @@
 const tmpl = require('Template/widget.pug');
-console.log(tmpl());
+
+console.log(tmpl({
+  data: {
+    items: {
+      a: 123,
+      b: 'abc'
+    }
+  }
+}));