+++ test/cases/require-assets-method-html/webpack.config.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const PugPlugin = require('pug-plugin');
+const PugPlugin = require('../../pug-plugin');
 
 module.exports = {
   mode: 'production',