+++ test/cases/require-assets-method-render/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const PugPlugin = require('pug-plugin');
+const PugPlugin = require('../../pug-plugin');
 
 module.exports = {
   mode: 'production',