+++ test/cases/require-resource-variable-sub-dir/webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
-const PugPlugin = require('pug-plugin');
-const basePath = path.resolve(__dirname);
+const PugPlugin = require('../../pug-plugin');
 
 module.exports = {
   stats: {
@@ -11,9 +10,6 @@ module.exports = {
 
   resolve: {
     alias: {
-      Source: path.join(basePath, 'src/'),
-      Includes: path.join(basePath, 'src/includes/'),
-      SourceImages: path.join(basePath, 'src/images/'),
     },
   },
 
@@ -23,7 +19,7 @@ module.exports = {
   },
 
   entry: {
-    index: 'src/includes/template/index.pug',
+    index: 'src/template/index.pug',
   },
 
   plugins: [new PugPlugin()],