+++ test/pug-plugin.js
@@ -1,0 +1,4 @@
+const PugPlugin = require('pug-plugin'); // npm package
+//const PugPlugin = require('../../pug-plugin'); // for local dev only
+
+module.exports = PugPlugin;